perf_test/requirements.txt
PyYAML==6.0.1
psutil==5.9.8
